feat(#235): 목표 타임라인 페이징 조회 API 작업

+++ backend/application/api/src/main/kotlin/io/raemian/api/emoji/model/EmojiCountSubset.kt
@@ -4,6 +4,6 @@ data class EmojiCountSubset(
     val id: Long,
     val name: String,
     val url: String,
-    val reactCount: Long,
+    val reactCount: Int,
     val isMyReaction: Boolean,
 )
